requirements.txt
git+https://github.com/huggingface/transformers
jsonlines
ftfy
lm_dataformat
numpy
pandas
tqdm
gdown
matplotlib
seaborn
black
datasets
bert_score
rouge_score
sentence-transformers
openai
ipykernel
jupyter